backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
chromadb
motor
sqlalchemy
pymysql
pandas
python-dotenv
